64 bit pit_t

+++ Kernel/include/types.h
@@ -19,7 +19,7 @@ typedef int64_t sem_t;
 // lock_t
 typedef int lock_t;
 
-typedef int16_t pid_t;
+typedef int64_t pid_t;
 
 typedef int fd_t;
 //
