+++ Userland/shell/include/syscalls.h
@@ -8,7 +8,7 @@
 // sem_t
 typedef int64_t sem_t;
 
-typedef int16_t pid_t;
+typedef int64_t pid_t;
 
 typedef int fd_t;
